+++ Classes/Renderer/AtomRenderer.php
@@ -18,6 +18,7 @@ use Brotkrueml\FeedGenerator\Contract\FeedInterface;
 use Brotkrueml\FeedGenerator\Contract\ImageInterface;
 use Brotkrueml\FeedGenerator\Contract\ItemInterface;
 use Brotkrueml\FeedGenerator\Contract\TextInterface;
+use Brotkrueml\FeedGenerator\Contract\XmlExtensionInterface;
 use Brotkrueml\FeedGenerator\Format\FeedFormat;
 
 /**
@@ -89,6 +90,9 @@ final class AtomRenderer implements RendererInterface
 
         foreach ($feed->getExtensionElements() as $element) {
             $extension = $this->extensionRegistry->getExtensionForElement(FeedFormat::ATOM, $element);
+            if (! $extension instanceof XmlExtensionInterface) {
+                continue;
+            }
             $extension->getXmlRenderer()->render($element, $root, $this->xml);
             $this->usedExtensions[$extension->getQualifiedName()] = $extension->getNamespace();
         }
@@ -215,6 +219,9 @@ final class AtomRenderer implements RendererInterface
 
         foreach ($item->getExtensionElements() as $element) {
             $extension = $this->extensionRegistry->getExtensionForElement(FeedFormat::ATOM, $element);
+            if (! $extension instanceof XmlExtensionInterface) {
+                continue;
+            }
             $extension->getXmlRenderer()->render($element, $itemNode, $this->xml);
             $this->usedExtensions[$extension->getQualifiedName()] = $extension->getNamespace();
         }