+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Extension API for Atom and RSS feeds
+- Extension API for feeds
 
 ## [0.5.0] - 2022-11-18